+++ drv/vsc7448/src/serdes1g.rs
@@ -42,7 +42,7 @@ impl Config {
             },
         }
     }
-    pub fn apply(&self, instance: u32, v: &Vsc7448Spi) -> Result<(), VscError> {
+    pub fn apply(&self, instance: u8, v: &Vsc7448Spi) -> Result<(), VscError> {
         v.serdes1g_read(instance)?;
         let ana_cfg = Vsc7448::HSIO().SERDES1G_ANA_CFG();
         let dig_cfg = Vsc7448::HSIO().SERDES1G_DIG_CFG();
